added vim

+++ bin/systemSetup.sh
@@ -12,7 +12,14 @@ function check_installation {
 	exit 1
 }
 
-BASIC_TOOLS=(unzip curl fc-cache alacritty tmux)
+BASIC_TOOLS=(unzip curl alacritty tmux nvim fzf docker docker-compose)
+if [ "$(uname -s)" == "Darwin" ]; then
+	BASIC_TOOLS+=(brew)
+else	
+	BASIC_TOOLS+=(fc-cache)
+fi
+
+
 for i in "${BASIC_TOOLS[@]}"
 do
 	check_installation "$i"
@@ -20,20 +27,25 @@ done
 echo "basic programs/tools are installes :)"
 
 # install nerdfont
-NERDFONT_TEST_FILE="JetBrains Mono Regular Nerd Font Complete Mono.ttf"
-NERDFONT_ZIP_FILE="JetBrainsMono.zip"
-NERDFONT_DOWNLOAD_LINK="https://github.com/ryanoasis/nerd-fonts/releases/download/v2.3.3/$NERDFONT_ZIP_FILE"
-LOCAL_FONT_DIR="$HOME/.local/share/fonts"
-[ ! -f "$LOCAL_FONT_DIR/$NERDFONT_TEST_FILE" ] && (
-	echo "installing nerdfonts"
-	mkdir -p "$LOCAL_FONT_DIR"
-	mkdir -p /tmp/installfonts
-	cd /tmp/installfonts
-	curl -L -o "$NERDFONT_ZIP_FILE" "$NERDFONT_DOWNLOAD_LINK"
-	unzip $NERDFONT_ZIP_FILE -d "$LOCAL_FONT_DIR"
-       	rm -rf /tmp/installfonts
-	fc-cache -f -v
-)
+if [ "$(uname -s)" == "Darwin" ]; then
+	brew tap homebrew/cask-fonts
+	brew install --cask font-jetbrains-mono-nerd-font
+else
+	NERDFONT_TEST_FILE="JetBrains Mono Regular Nerd Font Complete Mono.ttf"
+	NERDFONT_ZIP_FILE="JetBrainsMono.zip"
+	NERDFONT_DOWNLOAD_LINK="https://github.com/ryanoasis/nerd-fonts/releases/download/v2.3.3/$NERDFONT_ZIP_FILE"
+	LOCAL_FONT_DIR="$HOME/.local/share/fonts"
+	[ ! -f "$LOCAL_FONT_DIR/$NERDFONT_TEST_FILE" ] && (
+		echo "installing nerdfonts"
+		mkdir -p "$LOCAL_FONT_DIR"
+		mkdir -p /tmp/installfonts
+		cd /tmp/installfonts
+		curl -L -o "$NERDFONT_ZIP_FILE" "$NERDFONT_DOWNLOAD_LINK"
+		unzip $NERDFONT_ZIP_FILE -d "$LOCAL_FONT_DIR"
+       		rm -rf /tmp/installfonts
+		fc-cache -f -v
+	)
+fi
 
 # install alacritty themes
 ALACRITTY_THEMES_DIR="$HOME/.config/alacritty/themes"
